Documents/fastapikafka/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
kafka-python>=2.0.2
pydantic>=2.5.0
faker>=20.1.0
requests>=2.31.0
pymongo>=4.6.0
pyspark>=3.5.0
minio>=7.2.0
psycopg2-binary>=2.9.9
pandas>=2.0.0
jupyter>=1.0.0
notebook>=7.0.6
python-dotenv>=1.0.0 